AI-Powered-Form-Filling-Assistant-for-Indian-Citizen-Services-main/backend/requirements.txt
flask
pytesseract
opencv-python
pdf2image
pypdf
reportlab
Pillow
numpy
requests
python-multipart
# Optional (speech and NLP enhancements):
vosk
spacy
transformers
torch
